Handle errors on deleting a client
