+++ src/app/clients/getClientsData.ts
@@ -139,37 +139,3 @@ const getClientsCount = async (
 };
 
 export default getClientsDataAndCount;
-
-type GetClientIsActiveErrorType = "failedClientIsActiveFetch";
-export interface GetClientIsActiveError {
-    type: GetClientIsActiveErrorType;
-    logId: string;
-}
-
-type GetClientIsActiveResponse =
-    | {
-          error: null;
-          isActive: boolean;
-      }
-    | {
-          error: GetClientIsActiveError;
-          isActive: null;
-      };
-
-export const getIsClientActive = async (
-    supabase: Supabase,
-    clientId: string
-): Promise<GetClientIsActiveResponse> => {
-    const { data: isActiveData, error: isActiveError } = await supabase
-        .from("clients")
-        .select("primary_key, is_active")
-        .eq("primary_key", clientId)
-        .single();
-    if (isActiveError) {
-        const logId = await logErrorReturnLogId("Error with fetch: client table", {
-            error: isActiveError,
-        });
-        return { error: { type: "failedClientIsActiveFetch", logId }, isActive: null };
-    }
-    return { isActive: isActiveData.is_active, error: null };
-};
